+++ src/SMAPI.Web/Views/LogParser/Index.cshtml
@@ -1,6 +1,6 @@
 @using Newtonsoft.Json
+@using StardewModdingAPI.Web.Framework.ClientPlatformDetection
 @using StardewModdingAPI.Web.Framework.LogParsing.Models
-@using StardewModdingAPI.Web.Framework.UserAgentParsing
 @model StardewModdingAPI.Web.ViewModels.LogParserModel
 
 @{
